Add files via upload
The concept of travel and with special reference to small travel agencies, there is always a need to devise ways and means for improving customer satisfaction as well as efficiency of their operations. As a result of this, customer experiences during their travel can be highly enhanced by providing them with personalized travel planning, whereby their preferences will be taken into account in the formulation of their itinerary. This also makes it common to see people looking forward to being offered different specific trips, amounts and time suited to their requirements.